Added project structure
Added new folders
Started automated tests
